[CL-2084] Move ProjectFolderSelect outlet to main app
