+++ front/app/containers/Admin/projects/projectFolders/components/DeleteProjectFolderButton/StyledComponents.tsx
