+++ front/app/modules/commercial/admin_project_templates/admin/components/UseTemplateModal.tsx
@@ -18,7 +18,7 @@ import { client } from '../../utils/apolloUtils';
 import useAppConfigurationLocales from 'hooks/useAppConfigurationLocales';
 import useGraphqlTenantLocales from 'hooks/useGraphqlTenantLocales';
 import useAuthUser from 'hooks/useAuthUser';
-import { useProjectFolders } from 'modules/commercial/project_folders/hooks';
+import useProjectFolders from 'hooks/useProjectFolders';
 import { userModeratesFolder } from 'services/permissions/rules/projectFolderPermissions';
 import useLocalize from 'hooks/useLocalize';
