+++ front/app/modules/commercial/insights/admin/containers/Insights/List/CreateInsightsView.tsx
@@ -29,7 +29,7 @@ import GetProjects, { GetProjectsChildProps } from 'resources/GetProjects';
 import { PublicationStatus } from 'services/projects';
 // hooks
 import useLocalize from 'hooks/useLocalize';
-import useProjectFolders from 'modules/commercial/project_folders/hooks/useProjectFolders';
+import useProjectFolders from 'hooks/useProjectFolders';
 
 // services
 import { addInsightsView } from 'modules/commercial/insights/services/insightsViews';
