--- front/app/modules/commercial/project_folders/hooks/index.ts
@@ -1,2 +1,0 @@
-export { default as useProjectFolderModerators } from './useProjectFolderModerators';
-export { default as useProjectFolders } from './useProjectFolders';
